multiple npm versions (#23)
* use execSync and us it correctly
fix eslint error

* um...  global npm please, silly dev

* list the npm version for verification purposes

* Target all valid Developer node versions

We are targetting the node 4.3.2 Lambda environment for the deployed load generation code.  Make sure we target that environment.

However, it is valid for the developer's boxes to have any version of node greater than 4.  Therefore, we want to run the tests on each of those environments.