Change "run" command to be "invoke"
The command "invoke" is also more in line with the language used by Serverless.  Change to that from "run".
While at it, name options with short alias and give the full name as the alias.  This makes the documentation easier, with the short form listed first and then the long form.
Use "execute" (and it's forms) over "run" when talking about executing commands and code
Use "https" ubiquitously
Update tests
Clean up documentation
Remove NTP block.  This was done in a different commit that hasn't made it here yet.