Increase Hard-Coded Maximums
A number of parameters are used to determine when to split a script so that the entirety of the script can be executed on a single function.  This is helpful.  However, if you want to adjust them up, it is confining and requiring changing the code is largely a bit high of a bar.  Therefore, make the default values (max for single lambda) sensible (existing values) but make the hard coded maximum higher.  Do likewise for the script constraints.
Fix a few bugs:
   1. Missing validation for `maxChunkRequestsPerSecond`
   2. Missing validation for `timeBufferInMilliseconds`
   3. an incorrect comment.