no infinite loops, add process.exit feature
As per review comments, remove the infinite recursions in simulation and acceptance modes.

Add a feature that intercepts the result of the invocation from Serverless and uses that result to decide whether to exit the process with a non-zero exit code.  This is useful for "breaking" the build in a CI/CD pipeline use case.