Argument Pass Through
TL;DR: Change to using an argument pass-through strategy.  Resolves #7 and #72.  Add real unit tests.

Currently, we mostly emulate the command line arguments of the Serverless Framework (SLS).  As such, the argv re-assignment ugliness that we use can be eliminated, largely without impact.  This will eliminate whack-a-mole implementation requests for supporting argument `a` or `b` as time proceeds.  Further, it gives users a more coherent and predictable experience where users can apply their SLS knowledge (if it exists) to this new tool OR learn SLS light and graduate into using the full capabilities of SLS.

*THIS BREAKS DOWN*...
...if SLS ever chooses to allow the specification of the path to the serverless.yml file to execute (we currently run an algorithm to determine which `serverless.yml` file [local dir or global lib dir] to run SLS against).
...if SLS ever chooses to add a `-a` flag to the invoke command.  This would conflict with our "acceptance mode" flag.
...for the invoke command since we are, as a tool, adding assumptions about the meaning of the command.
The best example of this is the `-t` or `--type` flags; we programmatically determine whether we expect the given script to exceed the HTTP timeout used by SLS and choose the type appropriately.  As such, we must reject the use of this flag.
Another example is the `-f` parameter.  We assume a constant value for the function's name and as such must reject the supply of that flag.
A last example is the lack of providing a script is interpreted by this tool as an indication to use a local `script.yml`, falling back to the global `script.yml` if a local file with that name is not available.  SLS provides a "read from standard input" mode by default and so we will need to provide a means of regressing to that default mode in order to support such use cases.  In the same vein, if the user supplies the `-d`, `--data`, `--raw`, `-p`, or `--path` flags, we will need to take that as an indicator not to fill in the `-p` flag with the default script.

*BREAKING CHANGE*
We change the CLI.  `-s` (or `--script`) has a defined purpose with `sls invoke` and is interpretted as `stage` instead of `script` as we have defined it.  As such, all users will have to adjust to using the `-p` flag SLS uses to accept the script path.  While we're technically safe in making this change since we're still in 0.x.x, we don't want to abuse the relationship with our users.  For the short term, I'm searching for `-s` in the input and exiting with a failure to make sure we don't use the script as a stage name and invoke the serverless-artillery lambda with the default script (since `-p` will probably not be defined if they are supplying the `-s` flag).

We previous punted on providing thorough unit tests for the project.  These were a little hairy and disorganized so I have implemented many of the major test, particularly those describing behabvior implemented by this PR, and at least stubbed-out the tests for other parts of the code base for which tests seem appropriate.  *THESE WILL NEED TO BE IMPLEMENTED*  The nyc settings for allowable code coverage are abysmally low right now and should be improved in short order.