Fix Length is Duration Bug and Parse Failure
Fixes #83, #84, and #85.

The fix here:
https://github.com/Nordstrom/serverless-artillery/commit/7c1e56eac1de7729500f4746ed9401e775784521
missed the fact that the use of settings was the driver of the issue.  Code changes elsewhere fell prey to the missed naming=>expectation.

Our injection into the call SLS path can improperly alter the SLS stack calls if it does not properly handle parse failures.  There may be more to do here to ensure that a parse failure doesn't cause acceptance of a bad result.  However, the case where this came up was when invoking the function with invocation type "Event" and the payload is the empty string.  Therefore, the bug resolution is to both check for a non-empty string but also to protect the parse with a try catch in order to avoid future issues.

The handler uses > to compare script extent to limits, the CLI should not use >=