Length/Width => DurationInSeconds/RequestsPerSecond
Previously I had used a parallel set of terminology to describe script sizes.  This led to bugs and confusion.  As such, this commit removes all uses of the `length` and `width` monikers, preferring to use `DurationInSeconds` and `RequestsPerSecond` instead.

The hope is that this will reduce errors in the future and increase the clarity of the codebase.