Fix Duration Check Bug
The refactor of variable names was not properly validated against expected behavior.  As a result, the check for whether to use the RequestResponse invocation type instead of Event type fails to notice cases where the http timeout will be hit.  As a result, in those cases, the timeout is hit and the requester goes away, failing the lambda and causing retry (i.e. the load test runs twice).

Part of why this occurred is that we don't have a set of tests for script extent.  We recently vastly increased test coverage, this was something not tested.  I want this fix published before creating those tests.  The fix has been validated by a partner that was experiencing the issue.