Fix Duration Check Bug, Continued
First, the bleeding had to stop with a fix to get back to parity.

Now we need tests and perhaps a pattern that reduces ongoing bugs.  This PR provides those.

The logic for asking whether we need to use the event type invocation or request-response is simple.  We have an expected allowance, an expected consumption and somewhat conservative tolerances to account for a reasonable amount of unexpected.  This PR makes those assumptions more explicit and takes advantage of this to simplify the code where the results are used.

As per the prior PR, we needed tests to validate this feature.  These are added to a reasonable extent.  We can add more as we discover issues or change the implementation.  The refactor had an impact on the invoke implementation of course.

I twiddled just a bit with other tests to expand to hit otherwise missed defaults.