Refactor: runPerformance (separate splitting & execution)
handler.js:impl.runPerformance has been one of our terrible growing code burls.  This change refactors that.

The planning of an execution is separated out into a separate module (taskPlan) responsible for planning the splitting of the performance/load test specification into FaaS sized chunks.

The final execution of a task is separated out into a separate module (taskExec) responsible for behaving with feature parity as https://github.com/shoreditch-ops/artillery (though, to be honest, it does not yet - to be fixed)

These code movements implied the movement of tests of course.  The movements are incomplete but in a good interim state.  The continued movement of acceptance mode into the plan codebase follows and further removes code entanglement.  For clarity, that will be a separate commit.