Refactor: validScript (move script validation and mode declaration into a unifying task hive)
task.js: consolidate the task concerns to reduce requires and make the api more fluent.

handler.js:impl.validScript wove the task concerns into the handler.  instead, move it to the task to unweave the concerns.

taskValid.js + taskValid.spec.js: where the validation logic for scripts and modes definition moved to.

These code movements implied the movement of tests of course.  The movements are incomplete but in a good interim state.  The continued movement of acceptance mode into the plan codebase follows and further removes code entanglement.  For clarity, that will be a separate commit.