fix name bug
shortid can generate a name with an underscore within it ('_').  This character is invalid within the "service" attribute we were placing it into.  a quick, easy fix replaces all underscores with an 'A', this increases the appearance of 'A' but only very slightly increases the very low probability of collisions.

add test ensuring the replacement occurs.

replace some expect with should

fixes #96