Add `--jsonOnly`, Clean up logging (fix #100); add `npm run light`
Add the flags `-jo` and `--jsonOnly` to the invoke command that have the effect of suppressing all console output except for the JSON payload returned by the remote function.  This enables the piping of output to tools such as jq.

Use `log`, `warn`, and `error` as appropriate instead of just `log`
Clean up unintended logging and a commented out bit of logging that was for testing

Between these two changes, this commit fixes #100

For convenience, add a "light" script to the package.json that runs all tests excluding the long serverless-artillery.spec.js and npm.spec.js tests