Skip postinstall for SLS
We don't need tab completion with SLS in the context of our CI environment.  Therefore, skip the postinstall since it breaks the build.Skip postinstall for SLS  We don't need tab completion with SLS in the context of our CI environment.  Therefore, skip the postinstall since it breaks the build.