[WIP] Add Update Command
Add update command to CLI
Establish versioning folder to contain:
  1. old versions (for comparison)
  2. version specific update logic

Please note that the migration of logic from ~/lib/index.js into the versioning folder was entirely incomplete, as was most of the thinking associated with design and factoring of this feature.

Of particular note, this mostly implements the logic for taking an existing serverless.yml and non-destructively updating it with the monitoring mode assets.  Missing functionality included (as per #161) the updating of the package.json or of the handler's javascript files.  Update should be non-destructive, preserving any customizations that cannot be automatically transferred.