Adds upgrade command and implements for 0.0.0 to 0.0.1 (#194)
* Add Update Command

Add update command to CLI
Establish versioning folder to contain:
  1. old versions (for comparison)
  2. version specific update logic

Please note that the migration of logic from ~/lib/index.js into the versioning folder was entirely incomplete, as was most of the thinking associated with design and factoring of this feature.

Of particular note, this mostly implements the logic for taking an existing serverless.yml and non-destructively updating it with the monitoring mode assets.  Missing functionality included (as per #161) the updating of the package.json or of the handler's javascript files.  Update should be non-destructive, preserving any customizations that cannot be automatically transferred.

* Implements validation of project dependencies for upgrade command in versioning module.

* Adds validation of service definition.

* Unit tests all passing for validateServiceConfiguration (in upgrade context)  and validateServiceForDeployment.

* Renames asset directories by version number.
Makes local assets path a higher order parameter to versioning implementation.
Corrected update -> upgrade.
Relocate validateServiceForUpgrade function and test.

* Uses AJV and JSON schema definitions capture required preconditions and naming conflicts in the function service definition (serverless.yml).

* Implements the version by version upgrade by plugin mechanism.

* Removes now deprecated upgrade code from lib/index.js.

* Removes deprecated constants.

* Adds README to explain

* WIP: Upgrade saves to directories.

* Fix bugs preventing proper backup of project files and upgrade of package.json.
Establishes baseline versioning upgrade integration tests from 0.0.0 to 0.0.1.

* Repair some unit test drift.

* Refactor upgradeTest function to be more comprehensible.

* Fix typos and update versioning README to describe the revised backup mechanism.
Adds manual upgrade script for testing.

* Remove unnecessary debug logging.

* Adds output during upgrade to list files which were backed up.

* Minor adjustment to upgrade output.

* More verbose output explaining the next steps for users to finalize project upgrade. Improved output in general.